requirements.txt
Django==3.2
black==19.10b0
dj-database-url==0.5.0
django-cors-headers==3.4.0
django-extra-fields==3.0.2
django-filter==2.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
factory-boy==2.12.0
flake8==3.8.3
ipython==7.17.0
jedi==0.17.2
Pillow==8.2.0
psycopg2-binary==2.8.6
pytest==6.2.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-mock==3.3
python-decouple==3.3